pyproject.toml
[project]
name = "web-video-downloader"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.135.3",
    "playwright>=1.58.0",
    "pycryptodome>=3.23.0",
    "requests>=2.33.1",
    "tqdm>=4.67.3",
    "uvicorn>=0.44.0",
]
